fcoet: smatch errors
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_fc.c:306 fcoet_xfer_scsi_data() warn: right shifting more than type allows 8 vs 8
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_fc.c:523 fcoet_send_sol_abts() warn: right shifting more than type allows 8 vs 8
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_fc.c:802 fcoet_send_status() warn: right shifting more than type allows 16 vs 16
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_fc.c:802 fcoet_send_status() warn: right shifting more than type allows 16 vs 24
/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_fc.c:920 fcoet_disable_port() error: unchecked function return 'ss->ss_eport->eport_ctl'

/code/illumos-gate/usr/src/tools/proto/root_sparc-nd/opt/onbld/bin/sparc/smatch: ../../common/io/comstar/port/fcoet/fcoet_eth.c:1175 fcoet_process_sol_flogi_rsp() error: unchecked function return 'frm->frm_eport->eport_set_mac_address'

+++ usr/src/uts/common/io/comstar/port/fcoet/fcoet.h
@@ -207,7 +207,7 @@ typedef struct fcoet_exchange {
 
 	stmf_data_buf_t		**xch_dbufs;
 	uint8_t			xch_dbuf_num;
-	uint8_t			xch_sequence_no;
+	uint16_t		xch_sequence_no;
 	uint8_t			xch_ref;
 
 	int			 xch_left_data_size;
